requirements.txt
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
docopt==0.6.2
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
Js2Py==0.74
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pipwin==0.5.2
PyAudio==0.2.14
pygame==2.6.1
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
requests==2.32.3
setuptools==80.4.0
six==1.17.0
soupsieve==2.7
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
wheel==0.45.1
